common single loading test
