+++ library/src/test/java/io/reist/dali/main/MainSingleLoadingTest.java
@@ -1,10 +1,6 @@
 package io.reist.dali.main;
 
-import android.app.Activity;
-import android.graphics.Bitmap;
-import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.os.Bundle;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
@@ -16,14 +12,13 @@ import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricGradleTestRunner;
-import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
 
 import io.reist.dali.BuildConfig;
 import io.reist.dali.Dali;
-import io.reist.dali.DaliCallback;
 import io.reist.dali.DeferredImageLoader;
 import io.reist.dali.ShadowViewTreeObserver;
+import io.reist.dali.SingleLoadingTest;
 import io.reist.dali.TestImageView;
 
 /**
@@ -35,9 +30,8 @@ import io.reist.dali.TestImageView;
         sdk = {Build.VERSION_CODES.JELLY_BEAN},
         shadows = ShadowViewTreeObserver.class
 )
-public class MainSingleLoadingTest {
+public class MainSingleLoadingTest extends SingleLoadingTest {
 
-    public static final String TEST_URL = "test";
     public static final String TEST_URL_TRANSFORMED = TestImageTransformer.transformString(TEST_URL);
 
     @BeforeClass
@@ -47,17 +41,17 @@ public class MainSingleLoadingTest {
     }
 
     @Test
-    public void testLoadIntoImageView() {
-        ImageView imageView = Mockito.mock(ImageView.class);
-        Dali.load(TEST_URL).defer(false).into(imageView);
-        Mockito.verify(imageView).setImageDrawable(Mockito.any(Drawable.class));
-    }
+    public void testRequestTransformer() {
 
-    @Test
-    public void testLoadWithDaliCallback() {
-        DaliCallback daliCallback = Mockito.mock(DaliCallback.class);
-        Dali.load(TEST_URL).into(daliCallback, RuntimeEnvironment.application);
-        Mockito.verify(daliCallback).onImageLoaded(Mockito.any(Bitmap.class));
+        Dali.load(TEST_URL)
+                .defer(false)
+                .transformer(new TestImageTransformer())
+                .into(Mockito.mock(ImageView.class));
+
+        SyncTestImageLoader imageLoader = ((SyncTestImageLoader) Dali.getInstance().getMainImageLoader());
+
+        Assert.assertEquals(TEST_URL_TRANSFORMED, imageLoader.getUrl());
+
     }
 
     @Test
@@ -85,21 +79,7 @@ public class MainSingleLoadingTest {
 
     }
 
-    @Test
-    public void testRequestTransformer() {
-
-        Dali.load(TEST_URL)
-                .defer(false)
-                .transformer(new TestImageTransformer())
-                .into(Mockito.mock(ImageView.class));
-
-        SyncTestImageLoader imageLoader = ((SyncTestImageLoader) Dali.getInstance().getMainImageLoader());
-
-        Assert.assertEquals(TEST_URL_TRANSFORMED, imageLoader.getUrl());
-
-    }
-
-    public static void assertLoadingDeferred(View targetView, boolean shouldCallDefer) {
+    static void assertLoadingDeferred(View targetView, boolean shouldCallDefer) {
 
         Dali.load(TEST_URL).defer(true).into(targetView);
 
@@ -108,16 +88,6 @@ public class MainSingleLoadingTest {
                 targetView,
                 shouldCallDefer
         );
-
-    }
-
-    static class TestActivity extends Activity {
-
-        @Override
-        protected void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            setContentView(new ImageView(this));
-        }
 
     }
